Polish guides and editable markers demo for v0.16
