+++ demo-app/common/src/commonMain/kotlin/org/maplibre/compose/docsnippets/Location.kt
@@ -2,13 +2,18 @@
 
 package org.maplibre.compose.docsnippets
 
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import org.maplibre.compose.camera.CameraPosition
+import org.maplibre.compose.location.LocationPermission
 import org.maplibre.compose.location.LocationPuck
+import org.maplibre.compose.location.LocationState
 import org.maplibre.compose.location.LocationTrackingEffect
 import org.maplibre.compose.location.rememberDefaultHeadingProvider
 import org.maplibre.compose.location.rememberDefaultLocationProvider
 import org.maplibre.compose.location.rememberLocationState
+import org.maplibre.compose.location.rememberSystemSettingsLauncher
 import org.maplibre.compose.map.LocalMapState
 import org.maplibre.compose.map.MaplibreMap
 import org.maplibre.compose.map.rememberMapState
@@ -40,4 +45,35 @@ fun Location() {
   }
   MaplibreMap(state = mapState)
   // #endregion puck
+}
+
+@Composable
+private fun LocationPermissionButton(locationState: LocationState) {
+  // #region permission
+  if (locationState.permission !is LocationPermission.Granted) {
+    Button(onClick = locationState::requestPermission) {
+      Text("Use my location")
+    }
+  }
+  // #endregion permission
+}
+
+@Composable
+private fun LocationPermissionSettings(locationState: LocationState) {
+  // #region permission-settings
+  val settings = rememberSystemSettingsLauncher()
+  val permission = locationState.permission
+  if (permission is LocationPermission.NotGranted) {
+    when {
+      permission.shouldShowRationale -> {
+        Text("Your location helps you find places nearby.")
+        Button(onClick = locationState::requestPermission) { Text("Continue") }
+      }
+      permission.canRequest != false ->
+        Button(onClick = locationState::requestPermission) { Text("Use my location") }
+      settings.canOpenApplicationSettings ->
+        Button(onClick = { settings.openApplicationSettings() }) { Text("Open settings") }
+    }
+  }
+  // #endregion permission-settings
 }
