+++ docs/src/content/docs/expressions.mdx
@@ -53,23 +53,15 @@ features for which it is true:
 
 ## Missing data
 
-A feature property can be absent, so `feature["place"]` has the type
-`Expression<AnyValue?>`: the type is unknown and the value may be null. A
-nullable type argument follows Kotlin's rule. `Expression<StringValue?>` may
-evaluate to null and `Expression<StringValue>` may not.
-
-An assertion such as `asString()` aborts the expression for a feature without
-the property, and the map draws the property's default instead. Where a null is
-acceptable, keep the nullable type: a text or image property renders a null as
-nothing, and `eq`, `neq`, and `switch` on a value handle it. Where a value is
-required, supply one with <Api of="coalesce" /> and a fallback, or with the
-fallback argument of an assertion such as `asNumber(const(0f))`.
+Feature data can be incomplete. Use <Api of="coalesce" /> or an assertion's
+fallback argument, such as `asNumber(const(0f))`, to choose what to show when a
+value is missing:
 
 <Code code={rememberedMapStyle(region(expressions, "missing-data"))} lang="kotlin" title="App.kt" />
 
-<Api of="nil" /> is a `null` literal. MapLibre accepts it only where the
-expected type is unknown, such as in a comparison. To leave a layer property
-unset, pass Kotlin `null` to the layer.
+Use <Api of="nil" /> to compare an expression with null. To leave a layer
+property unset, pass Kotlin `null` to the layer. The API reference describes
+nullable expressions and type assertions in detail.
 
 ## Where the code runs
